finish using promise from render
