+++ react/typings/Session.d.ts
@@ -1,0 +1,21 @@
+interface SessionPickup {
+  name: string
+  address: SessionAddress
+}
+
+interface SessionData {
+  error?: any
+  response: {
+    id: string
+    namespaces: {
+      public?: {
+        favoritePickup?: {
+          value: {
+            name: string
+            address: SessionAddress
+          }
+        }
+      }
+    }
+  } | undefined
+}
